Corrected errors on timing computations on memcopies
git-svn-id: http://svn.code.sf.net/p/preesm/code/trunk@1619 c7aee7ca-e32f-b942-b52e-c1c222c4573e

+++ plugins/org.ietr.preesm.algorithm/src/org/ietr/preesm/algorithm/exportDif/DIFExporterVisitor.java
@@ -174,7 +174,7 @@ public class DIFExporterVisitor implements
 		// Retrieve execution time
 		// Only the execution runtime for cores with type x86 will be taken
 		// into account
-		int time = scenario.getTimingManager().getTimingOrDefault(
+		long time = scenario.getTimingManager().getTimingOrDefault(
 				vertex.getId(), "x86");
 		attributesMap.put("exec_time", time);
 		attributesMap.put("hierarchical", false);
